Rework equivalence checker proofs to be based on remove
This gets us almost all the way to the end of the proof.

The primary remaining issue is the fact that the current spec constrains
the registers holding the input and output values to not change across
the proof.  We need to rework how we're handling registers in the spec a
bit.